# openapi.yaml
openapi: 3.0.1
info:
  title: Digital signing service
  description: A service providing Kaleidos the functionality for digitally signing
    documents.
  license:
    name: MIT
    url: https://github.com/kanselarij-vlaanderen/digital-signing-service/LICENSE.txt
  version: 0.0.0
externalDocs:
  description: Git Repository
  url: https://github.com/kanselarij-vlaanderen/digital-signing-service
servers:
- url: /
paths:
  /sign-flows/{signflow_id}/signing/pieces:
    get:
      tags:
      - Signing files
      summary: List pieces associated with a sign flow
      operationId: listSignPieces
      parameters:
      - name: signflow_id
        in: path
        description: ID of sign flow
        required: true
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: successful operation
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/JsonApiData'
    post:
      tags:
      - Signing files
      summary: Prepare sign flow by uploading pieces to SigningHub
      operationId: addSignPieces
      parameters:
      - name: signflow_id
        in: path
        description: ID of sign flow
        required: true
        schema:
          type: string
          format: uuid
      responses:
        202:
          description: successful operation
  /sign-flows/{signflow_id}/signing/pieces/{piece_id}/signers:
    get:
      tags:
      - File signers
      summary: List signers for a certain file that is part of a publication flow
      operationId: listFileSigners
      parameters:
      - name: signflow_id
        in: path
        description: ID of sign flow
        required: true
        schema:
          type: string
          format: uuid
      - name: piece_id
        in: path
        description: ID of file
        required: true
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: successful operation
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/JsonApiData'
    post:
      tags:
      - File signers
      summary: Add signers for a certain file that is part of a publication flow
      operationId: addFileSigners
      parameters:
      - name: signflow_id
        in: path
        description: ID of sign flow
        required: true
        schema:
          type: string
          format: uuid
      - name: piece_id
        in: path
        description: ID of piece
        required: true
        schema:
          type: string
          format: uuid
      responses:
        202:
          description: successful operation
          content: {}
  /sign-flows/{signflow_id}/signing/pieces/{piece_id}/signinghub:
    get:
      tags:
      - Iframe link
      summary: Get an Iframe link for SigningHub interface integration
      operationId: getIframeLink
      parameters:
      - name: signflow_id
        in: path
        description: ID of sign flow
        required: true
        schema:
          type: string
          format: uuid
      - name: piece_id
        in: path
        description: ID of piece
        required: true
        schema:
          type: string
          format: uuid
      responses:
        303:
          description: Redirect to SigningHub page
  /sign-flows/{signflow_id}/signing/pieces/{piece_id}/start:
    post:
      tags:
      - Start signing
      summary: Send out file for signing by the appointed signers
      operationId: startSigning
      parameters:
      - name: signflow_id
        in: path
        description: ID of sign flow
        required: true
        schema:
          type: string
          format: uuid
      - name: piece_id
        in: path
        description: ID of piece
        required: true
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: Sucessfully sent out signing-flow
components:
  schemas:
    JsonApiResourceIdentification:
      type: object
      properties:
        type:
          type: string          
          enum:
          - pieces
          - mandatees
        id:
          type: string
          format: uuid
    JsonApiData:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/JsonApiResourceIdentification'